Griddlers:
This is a griddlers clone, in progress.  
right now you can create a new grid with it, including the number hints that actually make the puzzle.  
plans for the future:  
main menu  
play premade puzzles with mistake finder for hints    
brute force solver  
finding a more efficient way to solve the puzzle